requirements.txt
# ============================================================================
# SCOPONE - Project Dependencies
# ============================================================================
# Installa con: pip install -r requirements.txt
# ============================================================================

# GUI Framework
# tkinter è incluso con Python

# Modern UI Framework
customtkinter>=5.2.0

# Image Processing
Pillow>=10.0.0

# Optional: For more advanced features
# urllib3>=2.0.0  # For network features
# numpy>=1.24.0   # For data analysis
